+++ pvnet/data/__init__.py
@@ -1,1 +1,2 @@
 """Data parts"""
+from .utils import BatchSplitter
