+++ tests/data/test_datamodule.py
@@ -1,5 +1,6 @@
 from pvnet.data.datamodule import DataModule
-
+from pvnet.data.wind_datamodule import WindDataModule
+import os
 
 def test_init():
     dm = DataModule(
@@ -11,5 +12,18 @@ def test_init():
         val_period=[None, None],
         test_period=[None, None],
         block_nwp_and_sat=False,
+        batch_dir="tests/data/sample_batches",
+    )
+
+
+def test_wind_init():
+    dm = WindDataModule(
+        configuration=f"{os.path.dirname(os.path.abspath(__file__))}/data/sample_batches/data_configuration.yaml",
+        batch_size=2,
+        num_workers=0,
+        prefetch_factor=None,
+        train_period=[None, None],
+        val_period=[None, None],
+        test_period=[None, None],
         batch_dir="tests/data/sample_batches",
     )
